// libs/enums/doctor.enum.ts
export enum DoctorStatus {
	ACTIVE = 'ACTIVE',
	PENDING = 'PENDING',
	BLOCK = 'BLOCK',
	DELETE = 'DELETE',
}

export enum Specialization {
	UROLOGY = 'UROLOGY',
	DERMATOLOGY = 'DERMATOLOGY',
	PEDIATRICS = 'PEDIATRICS',
	SURGERY = 'SURGERY',
	DENTISTRY = 'DENTISTRY',
	CARDIOLOGY = 'CARDIOLOGY',
	ORTHOPEDICS = 'ORTHOPEDICS',
	OPHTHALMOLOGY = 'OPHTHALMOLOGY',
	OTHER = 'OTHER',
}